requirements.txt
aiohttp~=3.8.4
aiosqlite~=0.18.0
discord.py
google-cloud-compute